references/codebase-onboarding/projects/Manas2412__CodeBase-Q-A-with-RAG/pyproject.toml
# pyproject.toml
[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111"
uvicorn = {extras = ["standard"], version = "^0.30"}
sse-starlette = "^2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.29"
pgvector = "^0.3"
alembic = "^1.13"
voyageai = "^0.2"
anthropic = "^0.28"
cohere = "^5.5"
tree-sitter = "^0.22"
tree-sitter-python = "^0.22"
tree-sitter-javascript = "^0.22"
tree-sitter-typescript = "^0.22"
PyGithub = "^2.3"
gitpython = "^3.1"
celery = {extras = ["redis"], version = "^5.4"}
redis = {extras = ["asyncio"], version = "^5.0"}
ragas = "^0.1"
pydantic-settings = "^2.3"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
